LMS-16 - Connected the app to the mongodb database
