+++ package.json
@@ -5,8 +5,8 @@
   "main": "index.js",
   "scripts": {
     "build": "tsc",
-    "start": "node dist/server.js",
-    "dev": "tsx watch ./src/server.ts",
+    "start": "node dist/index.js",
+    "dev": "tsx watch ./src/index.ts",
     "lint": "eslint 'src/**/*.ts'",
     "lint:fix": "eslint 'src/**/*.ts' --fix"
   },
@@ -15,15 +15,20 @@
   "license": "ISC",
   "type": "commonjs",
   "devDependencies": {
+    "@types/cors": "^2.8.19",
     "@types/express": "^5.0.3",
     "@types/node": "^24.2.0",
-    "dotenv": "^17.2.1",
     "eslint": "^9.32.0",
-    "express": "^5.1.0",
     "install": "^0.13.0",
     "npm": "^11.5.2",
     "prettier": "^3.6.2",
     "ts-node": "^10.9.2",
     "typescript": "^5.9.2"
+  },
+  "dependencies": {
+    "dotenv": "^17.2.1",
+    "express": "^5.1.0",
+    "cors": "^2.8.5",
+    "mongoose": "^8.17.0"
   }
 }
