+++ src/config/config.ts
@@ -5,11 +5,13 @@ dotenv.config();
 interface Config {
     port: number;
     nodeEnv: string;
+    mongoUri: string;
 }
 
 const config: Config = {
     port: Number(process.env.PORT) || 3000,
     nodeEnv: process.env.NODE_ENV || 'development',
+    mongoUri: process.env.MONGO_URI || ''
 };
 
 export default config;
